Move `real_device` fixture to `conftest`
